feat(entity) New stategy to filter entities with a custom function.
