+++ packages/common/src/lib/entity/shared/strategies/index.ts
@@ -1,3 +1,3 @@
 export * from './strategy';
-export * from './filter-custom-state';
+export * from './filter-custom-function';
 export * from './filter-selection';
